refactor(client): implement paths corresponding FSD
