+++ packages/client/src/main/entity/room/model/index.ts
