+++ packages/client/src/main/runApp.tsx
@@ -1,7 +1,7 @@
 import { createRoot } from 'react-dom/client';
 import { StrictMode } from 'react';
 
-import { App } from './app';
+import { App } from 'app';
 
 /**
  * Run whole client application.
