+++ packages/client/tsconfig.eslint.json
@@ -2,6 +2,10 @@
   "extends": "../../tsconfig.eslint.json",
   "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.json", "./.eslintrc.js"],
   "compilerOptions": {
-    "types": ["vite/client"]
+    "types": ["vite/client"],
+    "baseUrl": ".",
+    "paths": {
+      "*": ["./src/main/*"]
+    }
   }
 }
